+++ CHANGELOG.md
@@ -11,7 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Players with [Player Locator Plus](https://modrinth.com/mod/player-locator-plus) or [Better Player Locator Bar](https://modrinth.com/mod/bplb) installed now require a compass item in the inventory to display the bar. This can be disabled in the config.
 
 ### Changed
+- Compat now has its own section in the config.
 - Xaero's Minimap compat can now be disabled.
+- No longer crashes on Forge and NeoForge servers.
 
 ## [1.0.1] - 04-20-2025